Replacing default `User` model with custom `HypUser`
This'll allow us to associate users and customers and make it possible
to further amend the data model for users in the future if we need to.

Generally we should try to not add things mindlessly to `HypUser`, and
instead break out separate models, but some enhancement from the default
seems likely to be needed and it's a pain to modify Django's default
user model once you have real production data.