Create customer when user signs up.
None of the customer/team/user management functionality will be built
out for MVP, but it'll be nice to have this part of the data model ready
to go.

In the future we can build out functionality for users to invite people
to their teams, manage permissions, deactivate people, etc.