Render error message if creating experiment fails.
Right now the error message is hardcoded because Django's built-in error
messages are not helpful for how we're rendering this form. Eg they say
things like "Your choice is not valid", which assumes we are rendering a
select box for choosing the experiment to associate with the variants.

Later we should add logic that intelligently generates error messages
for specific situations such as:
+ The experiment name is already taken.
+ The variant names for the experiment are not unique.