Fix uncertainty level test
1. Don't rely on Django ORM default sort order. Grabbing the `first()`
record in a set of associated records is apparently not the same as
getting the record at the 0-th index after calling `all()`.
2. Fix simulated variant conversion rates. Previously had a hashmap with
index keys and comparisons that evaluated once to become literal values
as the hashmap values. Needed to use the variants' IDs as keys and
desired conversion rate as values. Can then pass those conversion rates
to a function that evaluates the comparison function.