More flexible chart code
Decouple code for rendering charts from code specific to experiment
traffic splits (such as the chart title, DOM selectors, and object keys
needed to access the data).

This only took a few minutes to do an will both make it clearer what's
happening when you look at the template and make it easier to adapt the
charting code to other contexts.