Use `box-sizing: border-box;` for chart container
We want it to take up 100% of the width of the container, *including*
padding.

Should we use this setting globally for all elements? It seems
more intuitive to me when using percentage based dimensions.

However, it might be less intuitive for thinking about pixel or em based
dimensions, eg an element would be X pixels wide where Y of those pixels
are taken up by the padding or border.

This is different from the standard webkit interpretation of dimensions
which is width or height *plus* whatever padding and border have been
set.

Maybe it's better to stick with the default unless we need an exception
as we do in this case?

We could keep an eye out for other cases where we want to use
`box-sizing`, and make it a global style if more of those use cases
emerge.