Disambiguate `enqueue` functions
The `sample_data` module defines an `enqueue` function, but also imports
the `enqueue` function from the `job_utils` module.

Because of that Python was overwriting the `job_utils` definition with
the one in `sample_data`.

To resolve this we now explicitly use the `job_utils` module when
invoking its version of `enqueue`.